backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.0.3
python-multipart==0.0.6
matplotlib==3.7.2
seaborn==0.12.2
bcrypt==4.0.1
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

